Fix up the app benchmark scripts
In writing untar, realized large file was unavailable due to using
code-disk.img rather than disk.img. Fixed and added a check to
read_file.

+++ bench/concurrent/untar-and-search.sh
@@ -18,21 +18,29 @@ fuseopts=attr_timeout=0,entry_timeout=0,atomic_o_trunc
 if [ "$fs" = "cfscq" ]; then
   mnt="/tmp/fscq"
   cfscq disk.img /tmp/fscq +RTS "$rtsopt" -qg -RTS -- -o $fuseopts -f 1>/dev/null &
-  sleep 1
+  sleep 2
 fi
 
 if [ "$fs" = "fscq" ]; then
+  cp disk.img tmp-disk.img
   mnt="/tmp/fscq"
-  fscq disk.img /tmp/fscq +RTS "$rtsopt" -qg -RTS -- -o $fuseopts -f 1>/dev/null &
-  sleep 1
+  fscq tmp-disk.img /tmp/fscq +RTS "$rtsopt" -qg -RTS -- -o $fuseopts -f 1>/dev/null &
+  sleep 2
 fi
 
-read_file 1 /tmp/fscq/large >/dev/null
-rg 'PM_RESUME' /tmp/fscq/linux-4.10.11 >/dev/null
+if [ ! -d "$mnt/linux-4.10.11" ]; then
+  echo "could not find linux source"
+  exit 2
+fi
+
+read_file 1 "$mnt/large" >/dev/null
+rg 'PM_RESUME' "$mnt/linux-4.10.11" >/dev/null
 
 killall -SIGUSR1 $fs
 
-/usr/bin/time -f '%C\n %Uu %Ss %er' read_file 10 /tmp/fscq/large &
+cd "$mnt/linux-source"
+
+/usr/bin/time -f '%C\n %Uu %Ss %er' tar -xf linux.tar.xz &
 pid1="$!"
 /usr/bin/time -f '%C\n %Uu %Ss %er' rg -j3 'PM_RESUME' /tmp/fscq/linux-4.10.11 >/dev/null &
 pid2="$!"
